requirements.txt
# Python >=3.10 建議
flask
werkzeug
docling
requests
openai
gunicorn
python-dotenv
